+++ mycodo/scripts/upgrade_commands.sh
@@ -171,8 +171,8 @@ case "${1:-''}" in
     'update-influxdb')
         printf "\n#### Ensuring compatible version of influxdb is installed ####\n"
         INSTALL_ADDRESS="https://dl.influxdata.com/influxdb/releases/"
-        INSTALL_FILE="influxdb_1.3.7_armhf.deb"
-        CORRECT_VERSION="1.3.7-1"
+        INSTALL_FILE="influxdb_1.4.2_armhf.deb"
+        CORRECT_VERSION="1.4.2-1"
         CURRENT_VERSION=$(apt-cache policy influxdb | grep 'Installed' | gawk '{print $2}')
         if [ "${CURRENT_VERSION}" != "${CORRECT_VERSION}" ]; then
             echo "#### Incorrect InfluxDB version (v${CURRENT_VERSION}) installed. Installing v${CORRECT_VERSION}..."